+++ osu.Game/Screens/OnlinePlay/Multiplayer/MultiplayerMatchSongSelect.cs
@@ -2,23 +2,15 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using osu.Framework.Allocation;
-using osu.Framework.Bindables;
-using osu.Framework.Extensions.Color4Extensions;
-using osu.Framework.Graphics;
-using osu.Framework.Graphics.UserInterface;
 using osu.Framework.Logging;
 using osu.Framework.Screens;
-using osu.Game.Graphics;
 using osu.Game.Graphics.UserInterface;
 using osu.Game.Online.Multiplayer;
 using osu.Game.Online.Rooms;
 using osu.Game.Rulesets.Mods;
-using osu.Game.Screens.Play.HUD;
 using osu.Game.Screens.Select;
-using osuTK;
 
 namespace osu.Game.Screens.OnlinePlay.Multiplayer
 {
@@ -77,51 +69,5 @@ namespace osu.Game.Screens.OnlinePlay.Multiplayer
         protected override BeatmapDetailArea CreateBeatmapDetailArea() => new PlayBeatmapDetailArea();
 
         protected override bool IsValidFreeMod(Mod mod) => base.IsValidFreeMod(mod) && !(mod is ModTimeRamp) && !(mod is ModRateAdjust);
-    }
-
-    public class FooterButtonFreeMods : FooterButton, IHasCurrentValue<IReadOnlyList<Mod>>
-    {
-        public Bindable<IReadOnlyList<Mod>> Current
-        {
-            get => modDisplay.Current;
-            set => modDisplay.Current = value;
-        }
-
-        private readonly ModDisplay modDisplay;
-
-        public FooterButtonFreeMods()
-        {
-            ButtonContentContainer.Add(modDisplay = new ModDisplay
-            {
-                Anchor = Anchor.Centre,
-                Origin = Anchor.Centre,
-                DisplayUnrankedText = false,
-                Scale = new Vector2(0.8f),
-                ExpansionMode = ExpansionMode.AlwaysContracted,
-            });
-        }
-
-        [BackgroundDependencyLoader]
-        private void load(OsuColour colours)
-        {
-            SelectedColour = colours.Yellow;
-            DeselectedColour = SelectedColour.Opacity(0.5f);
-            Text = @"freemods";
-        }
-
-        protected override void LoadComplete()
-        {
-            base.LoadComplete();
-
-            Current.BindValueChanged(_ => updateModDisplay(), true);
-        }
-
-        private void updateModDisplay()
-        {
-            if (Current.Value?.Count > 0)
-                modDisplay.FadeIn();
-            else
-                modDisplay.FadeOut();
-        }
     }
 }
