+++ osu.Game/Screens/OnlinePlay/OnlinePlaySongSelect.cs
@@ -15,7 +15,7 @@ using osu.Game.Online.Rooms;
 using osu.Game.Overlays.Mods;
 using osu.Game.Rulesets;
 using osu.Game.Rulesets.Mods;
-using osu.Game.Screens.OnlinePlay.Multiplayer;
+using osu.Game.Screens.OnlinePlay.Match;
 using osu.Game.Screens.Select;
 
 namespace osu.Game.Screens.OnlinePlay
